requirements.txt
sqlalchemy
pydantic_settings
selenium
asyncpg
python-dotenv
telethon
cryptography 